Switch to legal moves
Pseudo-legal moves are best in a chess engine, but for a UI, it is more
practical to generate legal moves directly. Also, since performance does not
matter much here, we can remove pos->pieceOn[] and recompute as needed (ie.
sacrifice performance for robust DRY representation).

As a by-product, we can now count leaves in bulk at depth=1, which is 1-2 orders
of magnitude faster (factor = average number of pseudo-legal moves).