Position.fullMove
Some engines rely on it, so we can't just leave it as "1" just to avoid parsing
errors. For example, Stockfish deduces from the fullmove the gamePly, which it
uses in time management.

Also make the FEN parser more tolerant, so that rule50 and fullMove are optional
parameters with default values 0 and 1.