deadlines enforced by master thread
Workers spend most of their time waiting on blocking I/O reads. And these can
block for unbounded time periods, with buggy engines. So we must be 100% robust
to buggy engines. To achieve that, there is no other (good) choice than using a
differentthread (here the master thread) to check on progress.

This patch introduces the concept of deadlines, which are object shared between
workers and master thread, which periodically checks (and enforces) deadlines.

For now, only apply to the "go" command, but should really be applied to every
single I/O operation.