RAII for strings
One of the few good things about C++ is the ability to automatically set
destructors to clean up once a varialbe goes out of scope.

This is not possible in standard C, but fortunately we have GCC extensions, also
supported by Clang.

Verified with clang leak sanitizer (also memory sanitizer for good measure).