Redirect stderr in -debug mode
In normally runs, we don't need to read it, but in '-debug' mode, we want to
provide as much helpful information as possible to the programmer to debug the
engine. The typical use case, is when the engine crashes on an assert() which
writes to stderr. Of course, programmers can use stderr on an ad hoc basis as
well for logging certain things.