fix signedness bug
The bug was revealed by 6f634f5d, and otherwise resulted in silently broken code, doing
the random seek only half of the time, and a no-op otherwise.

The C library uses signed long for fseek. So we must be very careful when mixing signed
and unsigned:
- prng(&seed) % size: results in compiler warnings, but correct code.
- (long)prng(&seed) % size: misguided attempt to silence compiler warning, broken code,
  resulting in a negative result half of the time, hence a failed fseek().
- (long)(prng(&seed) % (uint64_t)size): is verbose, but the only correct solution.

Also, use some const str_t * for some of the high level API instead of const char *:
- advantage 1: easier to test for empty string (s.len=0 instead of s==NULL || *s=='\0').
- advantage 2: allows use of _s() faster functions that avoid strlen().
- disadvantage: can just call with C-string, especially string litteral for testing.