add chess960.epd
Useful from a user perspective, since c-chess-cli is not capable of auto generating the chess960 starting positions, but requires them to be fed by epd file.

Also useful to create an integration test. Example:

$ ./c-chess-cli -cmd ./test/engine -name engine1:engine2 -depth 8 -openings test/chess960.epd -chess960 -games 960 -pgnout out.pgn -debug > /dev/null
$ shasum out.pgn
83d32408f97033dd79edd2c5e6af8bfd9418614d  out.pgn

That's a pretty strong proof of functionality to verify patches from here on. To be automated and deployed via github CI.