implement a mini-shell parsing
Parse (with space disambiguation using \ delim rule) cmd into tokens, first one is the program to
run (with rules of extracting cwd applying), and the rest are arguments, which must be collected
and marshalled into a char ** for execvp().

Now c-chess-cli can run much more complicated commands, involving spaces in files and directories,
and program arguments. Need to document that, in a way somehow understandable by average users.