Fix double frees and leaks
This is where shit hits the fan:

for (size_t j = 0; j < vec_size(each.options); j++)
    vec_push((*eo)[i].options, each.options[j]);

The vec_push() is really doing a move operation, not a clone. That's because the each.options[j] are of type str_t, which merely contain a reference to the underlying memory. And we're doing several moves of the same string, due to the outer loop on i. So they get moved to different eo[i].options[j] array elements, and we end-up with multiple frees of the same memory once those eo[i].options[j] strings get freed.

Thankfully musl was here to save the day, and generated a segfault. On the other hand, glibc showed no symptoms, not even with debug compiles (perhaps multiple free()s are supposed to be a no-op in glibc).

Also, each.cmd and each.name were leaked at the end of the function. Here we use a robust RAII pattern, to make the code simple and maintainable.