fix dangerous anti-pattern
For RAII pattern to apply, only one constructor may be called during the lifetime of an object. Here each is constructed twice. First when each = {0}. Second when calling options_parse_eo() which is a constructor in disguise (allocates memory that is owned by the caller). Ownership should be clear: either of the functions must be the owner, not both. If each.cmd (for example) had been allocated somewhere in the code, then replacing it with each = options_parse_eo() would cause a memory leak.