pos_move_to_lan() should not own the output string
Effectively pos_move_to_lan() was a string constructor, which is a dangerous anti-pattern. Not only is it inefficient, leading to lots of malloc/free when used in tight loops. But also dangerous, because the result of the function should only be assigned to a variable declaration, and there is no way to enforce that. Assigning to a string with memory already allocated leads to a memory leak.