ban (str_t){0} footgun
Now a string is either valid (with memory allocated, even when s = "") or not. None of this in-between state with s = {0}, which is technically not a valid state, but still works thanks to cleverness (leveraging realloc(NULL, ...) and free(NULL) being no-op).

In hindsight, this is just a nest footguns, created gratuitiously in the name of cleverness, or to save some typing. Both of which are the usual wrong reasons to design footguns. Yes, {0} is less typing than str_new(). And yes, this means I can't use = {0} to initialize struct which contain str_t elements. But, that's good thing: it forces me to use robust RAII patterns (such struct should have an explicit constructor).

For example, testing for empty string using s.buf wasn't possible, because there are 2 cases (either s.buf is NULL or s.buf it not null, but s.buf[0] is '\0'). In turn that means passing s.buf to a function that accpets a const char * and was not written with the knowledge of this str_t footgun in mind is dangerous territory.