cleanup deadline logging
Remove the 'responded in time' message sent by deadline_overdue() called in the main thread. This is wrong. All we can say there, is that the main thread observed that a deadline is set, and not overdue yet. We don't know if the engine has responded yet (and it may well not have).