make c-chess-cli S-FEN agnostic
It is the user who supplied the EPD file, and therefore decides whether to use S-FEN or X-FEN notations. The problem is that we override this choice, by recomputing the FEN. And this is needed for position pruning feature in uci_position_command().

Removing position pruning, and just sending the FEN as it is in the EPD file would be the lazy solution. But:
- it's not as robust (parsing and rewriting the FEN cleans it from double spaces for example).
- position pruning makes life easier when debugging engines, giving much shorter move lists to reproduce bugs.
- it's also faster, although not mesurable in practice, except in artificial use cases.