simplify test/engine parser
Makes it more brittle as well, because it hinges on the assumption that multiple spaces between tokens (or trailing spaces) are forbidden. But that's also the point: such abuses are forbidden by the protocol, so we may as well have a test/engine that crashes, so as to detect regression in the UI when it starts to violate that rule.