delegate zombie purge to the kernel (Linux/Android only)
Normal engines terminate when c-chess-cli ends, by detecting EOF when reading from stdin (whose write-end was closed), or receiving SIGPIPE when writen to stdout (whose read-end was closed). But there are some special snowflakes out there, that somehow manage to survive and hog resources, in a loop that never checks for EOF when reading from stdin, nor ever writes to stdout. Cheng 4.39 is an example of such an engine.
Only Linux provides a simple and complete solution. So leave MacOSX out for now.

POSIX solution would be complicated...
- set signal handler in parent process main thread.
- block signals in each worker thread "immediately" when spawned.
- set pgid for engines to equal the parent's pgid, using getpgid() -before fork- and setpgid() -after fork in child-.
- signal handler calls kill(0, SIGTERM); sends SIGTERM to all processes with pgid of parent process (all engines and c-chess-cli itself).

... and yet still broken:
- some signals can't be handled (by design), such as SIGKILL.
- some really shouldn't be, such as SIGSEV.
- is it even possible to handle SIGTERM this way? If we throw SIGTERM to all process in the pgid, we throw it to ourself and enter a SIGTERM recursion...
- fraught with race conditions.