Revert "Add xboard support (#32)"
This reverts commit 92046e7533f070a6824c556e913f21685f9cc1a2.

After doing some research, and experimentation (writing the engine side), I came to the conclusion that CECP is too ugly to live. There is nothing wrong with the code from @sderr, which is fine. The problem lies firmly in the protocol. I don't want CECP support to spread like a cancer in c-chess-cli codebase. Inevitably, starting support will attract issues, which will add more and more code to deal with this hideous protocol. And there will be no end to this, given how this protocol "works".

In practice, all modern engines support UCI, save a few broken amateurish ones that nobody cares about. Even the old ones, that used to be CECP have migrated to UCI (sometimes keeping CECP support as well like Arasan).

This protocol should have been aborted 2 decades ago, and supporting today is a disservice to the community, because it only encourages more people to use it, which is harmful.