Increase isready timeout (#36)
When running on slow hardware such as a raspberry pi and using an engine that loads a NNUE binary as part of a setoption, the engine can timeout.  This was tested on a raspberry pi loading NNUE binaries in Minic.  Doubling the value resolves the issue for this use case.