Survive illegal FEN
Simply load the next one, and so on, until a valid one comes up. If all of them are illegal, that's fine, we'll just keep printing "illegal FEN" messages to stdout, until the user is bored and hits Ctrl+C.

Calling DIE() is not user-friendly. Indeed the user could have been running a tournament for hours unattended, and come back to see that it was aborted just because of an invalid FEN, somewhere in the middle of it.