Use vec prefix for vectors
These are not normal heap pointers, which can be passed to realloc() or free() directly. Storing a pointer for the element[0] instead of the header is merely a trick to leverage the type system, despite the lack of templates in C. It's best to make it clear which is a vector vs. normal pointer.